test(theme): measure tokens on the active line and selection
theme-contrast measured the editor surface only, which is why the
active line and the selection could sit below AA unnoticed. It now
resolves both layers, compositing the translucent ones over the
surface, and checks every token against all three.

Adds the other half of the trade. Contrast against a highlight can
always be satisfied by making it fade into the surface, which deletes
the feature to pass the test, so the active line and selection are
now pinned to a visible distance from the surface. Uses CIEDE2000
rather than a contrast ratio: these are large blocks of colour where
hue and chroma carry the signal.

Colour maths moved to a shared helper. Both suites parse the same CSS
and needed the same CIEDE2000, and a second copy of that formula would
drift from the first.

color-contrast comes off the axe accepted list, which drops dark from
12 accepted nodes to 6. Recorded there that axe only ever scans the
default amethyst palette; the other five are covered by the unit test.
